move auth context to a different file
