+++ src/components/Login/LoginPanel.tsx
@@ -18,8 +18,7 @@ import {
 } from 'firebase/auth'
 
 import LoginButton from '@/components/Login/LoginButton'
-
-import { useAuth } from '../../context/Auth'
+import { useAuth } from '@/context/Auth/context'
 
 export interface LoginPanelInterface {
   linkAccount: boolean // if true, will link the account to the logged in user, if false, will sign in with the provider
