+++ src/context/Auth/index.tsx
@@ -1,4 +1,4 @@
-import React, { createContext, useContext, useEffect, useState } from 'react'
+import React, { useEffect, useState } from 'react'
 import {
   Auth,
   AuthProvider,
@@ -11,23 +11,9 @@ import {
   User
 } from 'firebase/auth'
 
+import { AuthContext, FirebaseContextProps } from '@/context/Auth/context'
+
 import { auth } from '../../components/Firebase/init'
-
-interface FirebaseContextProps {
-  auth: Auth
-  getGoogleResults?: (auth: Auth) => void
-  linkAuthProvider?: (currentUser: User, provider: AuthProvider) => void
-  externalAuthSignIn?: (auth: Auth, provider: AuthProvider) => void
-  logOut?: () => void
-  currentUser: User | null
-}
-//set auth and current user as a context api to be called by other funcs
-const authContext = createContext<FirebaseContextProps>({
-  auth: auth,
-  currentUser: null
-})
-
-export const useAuth = () => useContext(authContext)
 
 export const AuthContextProvider = ({
   children
@@ -112,8 +98,8 @@ export const AuthContextProvider = ({
   }
 
   return (
-    <authContext.Provider value={value}>
+    <AuthContext.Provider value={value}>
       {!loading && children}
-    </authContext.Provider>
+    </AuthContext.Provider>
   )
 }
